:recycle: Extract each files
